ping360: Remove Ping360-V3.3.3_auto.hex
> We found another bug in the firmware and reached out to get a new version. We noticed that the head is not homing properly.
It should home after boot and any time the motor is turned off, but we noticed that this process was not happening.
This results in the sonar and ping viewer ending up out of sync so the image would be rotated compared to what the user would expect.
For now I would recommend falling back to 3.1.1 until we get the new version tested.

- Brian H

Signed-off-by: Patrick José Pereira <patrickelectric@gmail.com>